fix item unit factor

+++ app/Http/Requests/StoreItemRequest.php
@@ -32,8 +32,7 @@ class StoreItemRequest extends BaseRequest
             'item_category_id' =>'required|exists:item_categories,id|integer',
             'units' => ['array' , 'nullable' , 'max:3'],
             'units.*.name' => [  'required' , 'string'],
-            'units.*.conversion_factor'  => [ 'required' , 'integer' ],
-
+            'units.*.conversion_factor'  => [ 'required' , 'numeric' ],
 
         ];
     }
