+++ database/migrations/2022_03_21_235853_create_item_units_table.php
@@ -16,7 +16,7 @@ class CreateItemUnitsTable extends Migration
         Schema::create('item_units', function (Blueprint $table) {
             $table->id();
             $table->string('name');
-            $table->unsignedInteger('conversion_factor');
+            $table->float('conversion_factor');
             $table->boolean('is_const')->default(true);
             $table->boolean('is_default')->default(false);
             $table->foreignId('item_id')->constrained() ->onDelete('cascade');;
